requirements.txt
bs4
pysocks
requests
streamlit
python-dotenv
langchain-ollama
backports.tarfile
langchain_community
